Add endgame canvas
with menu, score, high score

+++ Assets/_Game/Scripts/Messaging/Messages/ScoreMessage.cs
@@ -5,10 +5,10 @@ namespace Messaging.Messages
 {
     public struct ScoreMessage
     {
-        public MathOperation Operation { get; }
-       public long Value { get; }
+       public MathOperation Operation { get; }
+       public int Value { get; }
 
-       public ScoreMessage(MathOperation op, long value = 0)
+       public ScoreMessage(MathOperation op, int value = 0)
        {
         Operation = op;
         Value = value;
